refactor: implement comprehensive code quality improvements
* Fix version synchronization in chocolatey installer (0.2.0 → 0.4.0)
* Remove reference to non-existent specs/ directory in README.md
* Refactor protocol.rs into specialized modules (worker_parser, task_parser, queue_parser)
* Consolidate 3 duplicate MockBroker implementations into shared test utilities
* Fix SSH security vulnerability in GitHub Actions release workflow
* Add comprehensive tidy analysis report with 165 improvement opportunities

All critical issues from tidy analysis have been resolved with full quality validation.
