+++ README.md
@@ -174,7 +174,4 @@ This runs formatting, linting, tests, and security audit.
 
 See [CONTRIBUTING.md](CONTRIBUTING.md) for guidelines.
 
-## Specifications
-
-See `specs/` directory for detailed specifications and planned features.
